# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tender-projects:
    get:
      operationId: listTenderProjects
      tags: [tenderProjects]
      summary: İhale projelerini getir
      responses:
        "200":
          description: İhale projeleri
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TenderProject"
    post:
      operationId: createTenderProject
      tags: [tenderProjects]
      summary: Yeni ihale projesi oluştur
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TenderProjectInput"
      responses:
        "201":
          description: Oluşturuldu
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenderProject"
        "400":
          description: Geçersiz istek
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /tender-projects/{id}:
    get:
      operationId: getTenderProject
      tags: [tenderProjects]
      summary: İhale projesi detayını getir
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: İhale projesi
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenderProject"
        "404":
          description: Bulunamadı
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteTenderProject
      tags: [tenderProjects]
      summary: İhale projesini sil
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Silindi
        "404":
          description: Bulunamadı
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /tender-projects/{id}/items:
    get:
      operationId: listTenderItems
      tags: [tenderProjects]
      summary: İhale projesindeki kalemleri getir
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: İhale kalemleri
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TenderItem"
        "404":
          description: Bulunamadı
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /tender-projects/{id}/upload:
    post:
      operationId: uploadTenderFile
      tags: [tenderProjects]
      summary: Excel formatındaki birim fiyat teklif cetvelini yükle ve satırları ayrıştır
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/FileUpload"
      responses:
        "200":
          description: Ayrıştırma sonucu
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenderUploadResult"
        "400":
          description: Dosya ayrıştırılamadı
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Bulunamadı
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /tender-projects/{id}/match:
    post:
      operationId: matchTenderProject
      tags: [tenderProjects]
      summary: Kalemleri seçilen birim fiyat listeleriyle otomatik eşleştir
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MatchRequest"
      responses:
        "200":
          description: Eşleştirme sonrası kalemler
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TenderItem"
        "404":
          description: Bulunamadı
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /tender-projects/{id}/export:
    get:
      operationId: exportTenderProject
      tags: [tenderProjects]
      summary: Doldurulmuş teklif cetvelini Excel olarak indir
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Excel dosyası
          content:
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary
        "404":
          description: Bulunamadı
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    FileUpload:
      type: object
      required: [file]
      properties:
        file:
          type: string
          format: binary
...
    TenderProject:
      type: object
      required:
        [
          id,
          name,
          fileName,
          itemCount,
          matchedCount,
          fuzzyCount,
          unmatchedCount,
          totalAmount,
          createdAt,
          updatedAt,
        ]
      properties:
        id:
          type: integer
        name:
          type: string
        fileName:
          type: ["string", "null"]
        itemCount:
          type: integer
        matchedCount:
          type: integer
        fuzzyCount:
          type: integer
        unmatchedCount:
          type: integer
        totalAmount:
          type: ["number", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    TenderProjectInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
    TenderUploadResult:
      type: object
      required: [project, items, warnings]
      properties:
        project:
          $ref: "#/components/schemas/TenderProject"
        items:
          type: array
          items:
            $ref: "#/components/schemas/TenderItem"
        warnings:
          type: array
          items:
            type: string
    TenderItem:
      type: object
      required:
        [
          id,
          tenderProjectId,
          rowOrder,
          pozNo,
          description,
          unit,
          quantity,
          unitPrice,
          totalPrice,
          matchStatus,
          matchScore,
          matchedPriceItemId,
          matchedPozNo,
          matchedDescription,
          matchedSourceName,
          createdAt,
        ]
      properties:
        id:
          type: integer
        tenderProjectId:
          type: integer
        rowOrder:
          type: integer
        pozNo:
          type: string
        description:
          type: string
        unit:
          type: ["string", "null"]
        quantity:
          type: number
        unitPrice:
          type: ["number", "null"]
        totalPrice:
          type: ["number", "null"]
        matchStatus:
          type: string
          enum: [matched, fuzzy, unmatched, manual]
        matchScore:
          type: ["number", "null"]
        matchedPriceItemId:
          type: ["integer", "null"]
        matchedPozNo:
          type: ["string", "null"]
        matchedDescription:
          type: ["string", "null"]
        matchedSourceName:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
...
    MatchRequest:
      type: object
      properties:
        priceListIds:
          type: ["array", "null"]
          items:
            type: integer
          description: Eşleştirme için kullanılacak liste id'leri. Boş/atlanırsa tüm listeler kullanılır.