# lib/api-zod/src/generated/types/matchRequest.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MatchRequest {
  /**
     * Eşleştirme için kullanılacak liste id'leri. Boş/atlanırsa tüm listeler kullanılır.
     * @nullable
     */
  priceListIds?: number[] | null;
}