# lib/api-zod/src/generated/types/tenderUploadResult.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TenderItem } from './tenderItem';
import type { TenderProject } from './tenderProject';

export interface TenderUploadResult {
  project: TenderProject;
  items: TenderItem[];
  warnings: string[];
}